+++ rusk-wallet/CHANGELOG.md
@@ -34,6 +34,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix contract call non-interactive parsing [#3602]
 - Fix exit when command errors in interactive mode [#3597], [#3192]
 - Fix out of order transaction history [#3702]
+- Fix empty transaction history error [#3700]
 
 ## [0.1.0] - 2025-01-20
 
@@ -70,6 +71,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 <!-- Issues -->
 [#3702]: https://github.com/dusk-network/rusk/issues/3702
 [#3707]: https://github.com/dusk-network/rusk/issues/3707
+[#3700]: https://github.com/dusk-network/rusk/issues/3700
 [#3597]: https://github.com/dusk-network/rusk/issues/3597
 [#3192]: https://github.com/dusk-network/rusk/issues/3192
 [#3650]: https://github.com/dusk-network/rusk/issues/3650